--- a/seclang/parser.go
+++ b/seclang/parser.go
@@ -10,7 +10,6 @@ import (
 	"io/fs"
 	"os"
 	"path/filepath"
-	"regexp"
 	"strings"
 
 	"github.com/corazawaf/coraza/v3"
@@ -77,40 +76,50 @@ func (p *Parser) FromFile(profilePath string) error {
 // or arguments are invalid
 func (p *Parser) FromString(data string) error {
 	scanner := bufio.NewScanner(strings.NewReader(data))
-	var linebuffer = ""
-	pattern := regexp.MustCompile(`\\(\s+)?$`)
+	var linebuffer strings.Builder
 	inQuotes := false
 	for scanner.Scan() {
 		p.currentLine++
 		line := strings.TrimSpace(scanner.Text())
-		if !inQuotes && len(line) > 0 && line[len(line)-1] == '`' {
+		lineLen := len(line)
+		if lineLen == 0 {
+			continue
+		}
+
+		if line[0] == '#' {
+			continue
+		}
+
+		if !inQuotes && line[lineLen-1] == '`' {
 			inQuotes = true
-		} else if inQuotes && len(line) > 0 && line[0] == '`' {
+		} else if inQuotes && line[0] == '`' {
 			inQuotes = false
 		}
+
 		if inQuotes {
-			linebuffer += line + "\n"
-		} else {
-			linebuffer += line
+			linebuffer.WriteString(line)
+			linebuffer.WriteString("\n")
+			continue
 		}
 
 		// Check if line ends with \
-		if !pattern.MatchString(line) && !inQuotes {
-			err := p.evaluate(linebuffer)
+		if line[lineLen-1] == '\\' {
+			linebuffer.WriteString(strings.TrimSuffix(line, "\\"))
+		} else {
+			linebuffer.WriteString(line)
+			err := p.evaluateLine(linebuffer.String())
 			if err != nil {
 				return err
 			}
-			linebuffer = ""
-		} else if !inQuotes {
-			linebuffer = strings.TrimSuffix(linebuffer, "\\")
+			linebuffer.Reset()
 		}
 	}
 	return nil
 }
 
-func (p *Parser) evaluate(data string) error {
+func (p *Parser) evaluateLine(data string) error {
 	if data == "" || data[0] == '#' {
-		return nil
+		return errors.New("invalid lines")
 	}
 	// first we get the directive
 	spl := strings.SplitN(data, " ", 2)
@@ -119,12 +128,11 @@ func (p *Parser) evaluate(data string) error {
 		opts = spl[1]
 	}
 	p.options.WAF.Logger.Debug("parsing directive %q", data)
-	directive := spl[0]
+	directive := strings.ToLower(spl[0])
 
 	if len(opts) >= 3 && opts[0] == '"' && opts[len(opts)-1] == '"' {
 		opts = strings.Trim(opts, `"`)
 	}
-	directive = strings.ToLower(directive)
 	if directive == "include" {
 		// this is a special hardcoded case
 		// we cannot add it as a directive type because there are recursion issues