This PR attempts to improve the parser performance.

Before:

```
➜  coraza git:(v3/improves_parser_performance) go test -benchmem -run=\^$ -bench \^BenchmarkParseFromString$ github.com/corazawaf/coraza/v3/seclang -benchtime=10s
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/seclang
BenchmarkParseFromString-8   	  894100	     11872 ns/op	   12525 B/op	     240 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/seclang	11.088s
```

After:

```
➜  coraza git:(v3/improves_parser_performance) go test -benchmem -run=\^$ -bench \^BenchmarkParseFromString$ github.com/corazawaf/coraza/v3/seclang -benchtime=10s
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/coraza/v3/seclang
BenchmarkParseFromString-8   	 1215052	      9866 ns/op	   10072 B/op	     216 allocs/op
PASS
ok  	github.com/corazawaf/coraza/v3/seclang	22.258s
```

> Thank you for contributing to Coraza WAF, your effort is greatly appreciated
> Before submitting check if what you want to add to `coraza` list meets [quality standards](https://github.com/corazawaf/coraza/blob/master/CONTRIBUTING.md#quality-standards) before sending pull request. Thanks!

**Make sure that you've checked the boxes below before you submit PR:**

- [ ] My code includes positive and negative tests.
- [ ] I have an appropriate description with correct grammar.
- [ ] I have read [Contribution guidelines](https://github.com/corazawaf/coraza/blob/master/CONTRIBUTING.md#contribution-guidelines), [maintainers note](https://github.com/corazawaf/coraza/blob/master/CONTRIBUTING.md#maintainers) and [Quality standard](https://github.com/github.com/corazawaf/coraza/v3sso/coraza-waf/blob/master/CONTRIBUTING.md#quality-standards).
- [ ] My code is properly linted and passes pre-commit tests.

Thanks for your contribution :heart: